third_party/blink/web_tests/http/tests/origin_trials/webexposed/app-history-origin-trial.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 AppHistory --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial" content="A+XNvNEjHaWhHcc4SQ0nNvrywJerpyfiZw5SEB7RLT/fnFZo5o1o7s+ORGAOi495CPCmfwuGAUpq0XB9TXcg/QUAAABSeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQXBwSGlzdG9yeSIsICJleHBpcnkiOiAyMDAwMDAwMDAwfQ==" />
<title>AppHistory API - interfaces and properties exposed by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(t => {
  assert_not_equals(window.appHistory, undefined);
  assert_not_equals(window.appHistory.current, null);
}, 'AppHistory API interfaces and properties in Origin-Trial enabled document.');
</script>
